Feat/atributos module

+++ backend/routes/api.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\Atributos\AtributosController;
 use App\Http\Controllers\Produtos\ProdutoController;
 use App\Http\Controllers\Auth\UserAuthController;
 use App\Http\Controllers\Caixa\CaixaController;
@@ -69,15 +70,24 @@ Route::prefix('empresas')->middleware(['auth:api', 'role:super_admin'])->group(f
 });
 
 
-Route::prefix('produtos')->middleware('auth:api', TenantResolver::class, TenantResolver::class)->group(function () {
+Route::prefix('produtos')->middleware('auth:api', TenantResolver::class)->group(function () {
     Route::get('/', [ProdutoController::class, 'index']);
     Route::get('{id}', [ProdutoController::class, 'show']);
     Route::get('/categoria/{id}', [ProdutoController::class, 'findByCategory']);
     Route::get('/barcode/{code}', [ProdutoController::class, 'findByBarcode']);
     Route::post('/', [ProdutoController::class, 'store']);
     Route::put('/{id}', [ProdutoController::class, 'update']);
-    Route::delete('{id}', [ProdutoController::class, 'delete']);
-    Route::delete('/', [ProdutoController::class, 'batchDelete']);
+    Route::delete('{id}', [ProdutoController::class, 'delete'])->middleware('role:admin|super_admin|gerente');
+    Route::delete('/', [ProdutoController::class, 'batchDelete'])->middleware('role:admin|super_admin|gerente');
+});
+
+Route::prefix('atributos')->middleware(['auth:api'])->group(function () {
+    Route::get('/', [AtributosController::class, 'index']);
+    Route::get('/por-tipo-de-negocio', [AtributosController::class, 'indexByTipoDeNegocio']);
+    Route::post('/', [AtributosController::class, 'store'])->middleware('role:admin|super_admin|gerente');
+    Route::get('{id}', [AtributosController::class, 'show'])->middleware('role:admin|super_admin|gerente');
+    Route::put('{id}', [AtributosController::class, 'update'])->middleware('role:admin|super_admin|gerente');
+    Route::delete('{id}', [AtributosController::class, 'delete'])->middleware('role:admin|super_admin|gerente');
 });
 
 // Crie uma rota de teste temporária:
@@ -122,14 +132,14 @@ JsonApiRoute::server('v1')
                     ]);
     });
 
-Route::prefix('relatorios')->middleware(['auth:api', 'role:admin|super_admin'], TenantResolver::class)->group(function () {
+Route::prefix('relatorios')->middleware(['auth:api'], TenantResolver::class)->group(function () {
     Route::get('/resumo', [RelatoriosController::class, 'getSalesSummary']);
-    Route::get('/por-categoria', [RelatoriosController::class, 'getSalesByCategory']);
-    Route::get('/produtos-mais-vendidos', [RelatoriosController::class, 'getTopProducts']);
-    Route::get('/receita-por-periodo', [RelatoriosController::class, 'getRevenueByPeriod']);
+    Route::get('/por-categoria', [RelatoriosController::class, 'getSalesByCategory'])->middleware('role:admin|super_admin');;
+    Route::get('/produtos-mais-vendidos', [RelatoriosController::class, 'getTopProducts'])->middleware('role:admin|super_admin');;
+    Route::get('/receita-por-periodo', [RelatoriosController::class, 'getRevenueByPeriod'])->middleware('role:admin|super_admin');;
     Route::get('/pedidos', [RelatoriosController::class, 'getOrders']);
     Route::prefix('comissoes')->group(function () {
-        Route::get('/', [CommissionsController::class, 'index']); // Comissões do mês atual
+        Route::get('/', [CommissionsController::class, 'index'])->middleware('role:admin|super_admin'); // Comissões do mês atual
         Route::get('/vendedor/{id}', [CommissionsController::class, 'getCommissionsByVendedor']);
         Route::get('/vendedor/{id}/comissoes', [CommissionsController::class, 'getCommissionsByVendedorAndDate']);
     });
@@ -138,9 +148,9 @@ Route::prefix('relatorios')->middleware(['auth:api', 'role:admin|super_admin'], 
 Route::prefix('categorias')->middleware('auth:api', TenantResolver::class)->group(function () {
     Route::get('/', [CategoriasController::class, 'index']);
     Route::get('{id}', [CategoriasController::class, 'show']);
-    Route::post('/', [CategoriasController::class, 'store']);
-    Route::put('{id}', [CategoriasController::class, 'update']);
-    Route::delete('{id}', [CategoriasController::class, 'delete']);
+    Route::post('/', [CategoriasController::class, 'store'])->middleware('role:admin|super_admin|gerente');
+    Route::put('{id}', [CategoriasController::class, 'update'])->middleware('role:admin|super_admin|gerente');
+    Route::delete('{id}', [CategoriasController::class, 'delete'])->middleware('role:admin|super_admin|gerente');
 });
 
 Route::prefix('clientes')->middleware('auth:api', TenantResolver::class)->group(function () {
@@ -148,7 +158,7 @@ Route::prefix('clientes')->middleware('auth:api', TenantResolver::class)->group(
     Route::get('{id}', [ClientesController::class, 'show']);
     Route::post('/', [ClientesController::class, 'store']);
     Route::put('{id}', [ClientesController::class, 'update']);
-    Route::delete('{id}', [ClientesController::class, 'delete']);
+    Route::delete('{id}', [ClientesController::class, 'delete'])->middleware('role:admin|super_admin|gerente');;
 });
 
 Route::prefix('variantes')->middleware('auth:api', TenantResolver::class)->group(function () {
@@ -156,7 +166,7 @@ Route::prefix('variantes')->middleware('auth:api', TenantResolver::class)->group
     Route::get('{id}', [VariantesController::class, 'show']);
     Route::post('/', [VariantesController::class, 'store']);
     Route::put('{id}', [VariantesController::class, 'update']);
-    Route::delete('{id}', [VariantesController::class, 'delete']);
+    Route::delete('{id}', [VariantesController::class, 'delete'])->middleware('role:admin|super_admin|gerente');;
 });
 
 Route::middleware('auth:api', TenantResolver::class)->group(function () {
@@ -177,34 +187,34 @@ Route::middleware('auth:api', TenantResolver::class)->group(function () {
 Route::middleware('auth:api', TenantResolver::class)->group(function () {
     Route::get('payment-methods', [PaymentMethodController::class, 'index']);
     Route::get('payment-methods/{id}', [PaymentMethodController::class, 'show']);
-    Route::post('payment-methods', [PaymentMethodController::class, 'store']);
-    Route::put('payment-methods/{id}', [PaymentMethodController::class, 'update']);
-    Route::delete('payment-methods/{id}', [PaymentMethodController::class, 'destroy']); 
+    Route::post('payment-methods', [PaymentMethodController::class, 'store'])->middleware('role:admin|super_admin|gerente');;
+    Route::put('payment-methods/{id}', [PaymentMethodController::class, 'update'])->middleware('role:admin|super_admin|gerente');;
+    Route::delete('payment-methods/{id}', [PaymentMethodController::class, 'destroy'])->middleware('role:admin|super_admin|gerente');; 
 }   
 );
 // Rotas para Tipos de Produtos
 Route::prefix('tipos-produtos')->middleware('auth:api', TenantResolver::class)->group(function () {
     Route::get('/', [TiposDeProdutosController::class, 'index']);
     Route::get('{id}', [TiposDeProdutosController::class, 'show']);
-    Route::post('/', [TiposDeProdutosController::class, 'store']);
-    Route::put('{id}', [TiposDeProdutosController::class, 'update']);
-    Route::delete('{id}', [TiposDeProdutosController::class, 'destroy']);
+    Route::post('/', [TiposDeProdutosController::class, 'store'])->middleware('role:admin|super_admin|gerente');;
+    Route::put('{id}', [TiposDeProdutosController::class, 'update'])->middleware('role:admin|super_admin|gerente');;
+    Route::delete('{id}', [TiposDeProdutosController::class, 'destroy'])->middleware('role:admin|super_admin|gerente');;
 });
 
 // Rotas para Tipos de Negócios
 Route::prefix('tipos-negocios')->middleware('auth:api', TenantResolver::class)->group(function () {
     Route::get('/', [TipoDeNegociosController::class, 'index']);
     Route::get('{id}', [TipoDeNegociosController::class, 'show']);
-    Route::post('/', [TipoDeNegociosController::class, 'store']);
-    Route::put('{id}', [TipoDeNegociosController::class, 'update']);
-    Route::delete('{id}', [TipoDeNegociosController::class, 'destroy']);
+    Route::post('/', [TipoDeNegociosController::class, 'store'])->middleware('role:admin|super_admin|gerente');;
+    Route::put('{id}', [TipoDeNegociosController::class, 'update'])->middleware('role:admin|super_admin|gerente');;
+    Route::delete('{id}', [TipoDeNegociosController::class, 'destroy'])->middleware('role:admin|super_admin|gerente');;
 });
 
 // Rotas para Configurações de Negócio
 Route::prefix('config-negocio')->middleware('auth:api', TenantResolver::class)->group(function () {
     Route::get('/', [ConfigDoNegocioController::class, 'index']);
     Route::get('{id}', [ConfigDoNegocioController::class, 'show']);
-    Route::post('/', [ConfigDoNegocioController::class, 'store']);
-    Route::put('{id}', [ConfigDoNegocioController::class, 'update']);
-    Route::delete('{id}', [ConfigDoNegocioController::class, 'destroy']);
+    Route::post('/', [ConfigDoNegocioController::class, 'store'])->middleware('role:admin|super_admin|gerente');;
+    Route::put('{id}', [ConfigDoNegocioController::class, 'update'])->middleware('role:admin|super_admin|gerente');;
+    Route::delete('{id}', [ConfigDoNegocioController::class, 'destroy'])->middleware('role:admin|super_admin|gerente');;
 });
